refactor : pr 반영) null 체크 삭제 및 코드 포맷
